tooloo_v4_hub/organs/memory_organ/requirements.txt
google-cloud-firestore
google-cloud-aiplatform
uvicorn
fastapi
pydantic
asyncio
numpy
mcp[server]
